+++ test/lifecycle/map-response.test.ts
@@ -51,6 +51,27 @@ describe('Map Response', () => {
 		expect(headers.get('X-Powered-By')).toContain('Elysia')
 	})
 
+	it('inherits plugin', async () => {
+		const plugin = new Elysia().mapResponse(
+			{ as: 'global' },
+			() => new Response('Fubuki')
+		)
+
+		const app = new Elysia().use(plugin).get('/', () => 'a')
+
+		const res = await app.handle(req('/')).then((t) => t.text())
+		expect(res).toBe('Fubuki')
+	})
+
+	it('not inherits plugin on local', async () => {
+		const plugin = new Elysia().mapResponse(() => new Response('Fubuki'))
+
+		const app = new Elysia().use(plugin).get('/', () => 'a')
+
+		const res = await app.handle(req('/')).then((t) => t.text())
+		expect(res).toBe('a')
+	})
+
 	it('map response only once', async () => {
 		const app = new Elysia().get('/', () => 'Hutao', {
 			mapResponse: [
@@ -120,6 +141,23 @@ describe('Map Response', () => {
 		const res = await app.handle(req('/')).then((x) => x.text())
 
 		expect(res).toBe('Hutao')
+	})
+
+	it('map response in order', async () => {
+		let order = <string[]>[]
+
+		const app = new Elysia()
+			.mapResponse(() => {
+				order.push('A')
+			})
+			.mapResponse(() => {
+				order.push('B')
+			})
+			.get('/', () => '')
+
+		await app.handle(req('/'))
+
+		expect(order).toEqual(['A', 'B'])
 	})
 
 	it('as global', async () => {